Refactor namespaces to improve organization and clarity in data models
