+++ src/TB.DanceDance.Mobile/Platforms/Android/UploadForegroundService.cs
@@ -3,6 +3,7 @@ using Android.Content;
 using Android.OS;
 using Android.Util;
 using TB.DanceDance.Mobile.Data;
+using TB.DanceDance.Mobile.Data.Models.Storage;
 using TB.DanceDance.Mobile.Services.DanceApi;
 using Debug = System.Diagnostics.Debug;
